Correct Digger behaviour

+++ src/main/java/dev/phomc/grimoire/enchantment/tool/DiggerEnchantment.java
@@ -18,7 +18,7 @@ public class DiggerEnchantment extends GrimoireEnchantment {
         super(identifier, EnchantmentRarity.RARE, EnchantmentTarget.PICKAXE);
     }
 
-    public void dig(ServerPlayer player, BlockPos origin, BlockState blockState) {
+    public void dig(ServerPlayer player, BlockPos origin, BlockState originState) {
         // ignoredPlayers prevents infinite recursion
         if (((ServerPlayerAccessor) player).shouldIgnoreDigger()) return;
         ((ServerPlayerAccessor) player).ignoreDigger(true);
@@ -27,11 +27,18 @@ public class DiggerEnchantment extends GrimoireEnchantment {
             for (int y = -RADIUS; y <= RADIUS; y++) {
                 for (int z = -RADIUS; z <= RADIUS; z++) {
                     BlockPos p = origin.offset(x, y, z);
+                    BlockState blockState = player.level.getBlockState(p);
                     // must not be origin (it has already broken)
                     // must not in blacklist
+                    // the destruction time must not exceed the origin's
+                    // the block can be interacted (with spawn check later)
                     // must have correct tool used
-                    if (p != origin && !blockState.is(GrimoireBlockTags.DIGGER_BLACKLIST) && player.hasCorrectToolForDrops(blockState)) {
-                        // this handles the rest (exp, drops, sound, item durability, etc)
+                    if (p != origin && !blockState.is(GrimoireBlockTags.DIGGER_BLACKLIST) &&
+                            blockState.getBlock().defaultDestroyTime() <= originState.getBlock().defaultDestroyTime() &&
+                            player.level.mayInteract(player, p) &&
+                            player.hasCorrectToolForDrops(blockState)) {
+                        // this handles the rest: do extra validation
+                        // and calculate exp, drops, sound, item durability, etc
                         player.gameMode.destroyBlock(p);
                     }
                 }
